updated telemetry
